Ammended JayZ.java back to Duke.java to ensure that the grading script runs smoothly
